image_proxy/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1google-genai==1.2.0
